Move most of ConstantResovler to zeitwerk_utils
