+++ Cargo.toml
@@ -55,3 +55,4 @@ predicates = "3.0.2" # kind of like rspec assertions
 # see #[serial] tag for more info
 # more info: https://fdeantoni.medium.com/running-tests-sequentially-in-rust-eed7566f63f0
 serial_test = "2.0.0"
+pretty_assertions = "1.3.0" # Shows a more readable diff when comparing objects
